changes to details page to read and write alll values properly. pls see the details.ts file for the format of the buffer written to the device.
Ref Prithvi: we have decided to use only 2 chars. a UUID_SENSE_BOARD_SETTINGS which will read
things like firmware version, board revision/manuf details etc.
and a UUID_SENSE_PI_USER_SETTINGS to write all the users settings to 1 charateristic. (11 bytes as of now)
(the format/order of the bytes specified in details.ts)

eventually we shd check the settings we read and then write if compatible. as of now, not reading.
and even more eventually :) we shd actually determine what settings can be modified by reading the settings from
the board.

current issues:
1. i get an error 13 when writing : could be an issue on the board side : not letting me write 11 bytes. to be checked.
2. the scannign takes much too long. even now, sometimes 1-2 minutes before the sensepi is discovered by the app. check the scan settings