re-architecting, work in progress version of setup : quick summary from my msg to prithvi
we need to get in place things like reading the name/services etc properly from the scan page,
so we can load config for the appropriate device (sensepi or sensebe, or ac).
also the code needs to use formbuilders so we can do better validation at the UI itself.
there are things to understand in the ionic world like how to create "libraries" for each sense type :
we can use something called providers,, but i need to understand how to also bind the UI with
the data in each library/provider, along with custom validators for each.
so working through these arch type details in ionic. once this backend is in place, in a modular,
systematic way, i think deisigning the front end to breakup settings across mulitple windows
etc will not be very complex.
right now there is such tight binding between the frontend and backend, its hard/imposs to
change anything for new devices/types/versions etc