added the formvalidation code. so its now easy to write our own validation functions for every input, or enforce limits on user input (eg: only single decimal nums between 1 and 500 etc)
there appears to be some porblem with the ngIf code. it is not hiding/showing divs based on the boolean. :(

one hack to trigger the cards is to enter a value in the timer interval which is always displayed (it shd not be)
and then the rest of the cards obey the ngif.
why?

TBD:
0. fix the ngif bug :(
1.now display the data in slides/different screens etc. instead of just scrolling
2. the number keyboard appears to occupy too much of the screen. find a different one?