1. added a default landing page for sensebe. if we find a device with our UUID (right now using a sensepi uuid, add sensebe) and a shortened name that starts with SB, it will load a sensebe config page on connection. if SP is the starting str, it loads a sensepi page
BUG:
found a scenario where the conditional ion-cards work.
the cards are all loaded based on whether the user selects timer or pir mode
we use an ngif to control the view of these cards
however, when the user changes his selection, timer or pir, the cards shown do not change, UNLESS, we change the orientation of the device.
when orientation is changed, all the cards are refreshed properly.
am missing some setting here. dont know yet, why its not just refreshing the cards on change in selection.